Refactor import statements for scanner utilities
